Fixing northeast/northwest/southeast/southwest directions in import rm2k charsets thing
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@1144 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/common/Image32.cpp
@@ -689,9 +689,9 @@ inline void copyAlpha(RGBA& dest, RGBA src)
 
 inline void additiveRGBA(RGBA& dest, RGBA src)
 {
-  dest.red   = std::max(0, std::min(255, (dest.red   + ((src.red   * src.alpha) / 255))));
-  dest.green = std::max(0, std::min(255, (dest.green + ((src.green * src.alpha) / 255))));
-  dest.blue  = std::max(0, std::min(255, (dest.blue  + ((src.blue  * src.alpha) / 255))));
+  dest.red   = std::max(0, std::min(255, (dest.red   + (((int)src.red   * (int)src.alpha) / 255))));
+  dest.green = std::max(0, std::min(255, (dest.green + (((int)src.green * (int)src.alpha) / 255))));
+  dest.blue  = std::max(0, std::min(255, (dest.blue  + (((int)src.blue  * (int)src.alpha) / 255))));
 }
 
 ////////////////////////////////////////////////////////////////////////////////
